+++ src/gromacs/mdlib/nbnxn_atomdata.h
@@ -1,7 +1,7 @@
 /*
  * This file is part of the GROMACS molecular simulation package.
  *
- * Copyright (c) 2012,2013,2014,2015,2016,2017,2018, by the GROMACS development team, led by
+ * Copyright (c) 2012,2013,2014,2015,2016,2017,2018,2019, by the GROMACS development team, led by
  * Mark Abraham, David van der Spoel, Berk Hess, and Erik Lindahl,
  * and including many others, as listed in the AUTHORS file in the
  * top-level source directory and at http://www.gromacs.org.
@@ -61,7 +61,7 @@ void nbnxn_free_aligned(void *ptr);
 
 /* Reallocation wrapper function for nbnxn data structures */
 void nbnxn_realloc_void(void **ptr,
-                        int nbytes_copy, int nbytes_new,
+                        size_t nbytes_copy, size_t nbytes_new,
                         nbnxn_alloc_t *ma,
                         nbnxn_free_t  *mf);
